Sync

+++ sd/source/ui/remotecontrol/ImagePreparer.cxx
@@ -24,19 +24,10 @@
 #include <comphelper/processfactory.hxx>
 #include <osl/file.hxx>
 #include <rtl/ustrbuf.hxx>
-#include <sax/tools/converter.hxx>
 #include <rtl/strbuf.hxx>
 #include <sal/log.hxx>
-#include <unotools/streamwrap.hxx>
-
-#include <svl/itemset.hxx>
-#include <sfx2/docfile.hxx>
 
 #include <com/sun/star/beans/PropertyValue.hpp>
-#include <com/sun/star/container/XNameAccess.hpp>
-#include <com/sun/star/document/XFilter.hpp>
-#include <com/sun/star/document/XImporter.hpp>
-#include <com/sun/star/document/XExporter.hpp>
 #include <com/sun/star/drawing/GraphicExportFilter.hpp>
 #include <com/sun/star/lang/XServiceName.hpp>
 #include <com/sun/star/presentation/XSlideShowController.hpp>
@@ -155,16 +146,19 @@ uno::Sequence<sal_Int8> ImagePreparer::preparePreview(
 
     xFilter->filter( aProps );
 
-    // FIXME: error handling.
+    File aFile(aFileURL);
+    if (aFile.open(0) != osl::File::E_None)
+        return uno::Sequence<sal_Int8>();
 
-    File aFile( aFileURL );
-    aFile.open(0);
     sal_uInt64 aRead;
     rSize = 0;
     aFile.getSize( rSize );
     uno::Sequence<sal_Int8> aContents( rSize );
 
     aFile.read( aContents.getArray(), rSize, aRead );
+    if (aRead != rSize)
+        aContents.realloc(aRead);
+
     aFile.close();
     File::remove( aFileURL );
     return aContents;
@@ -183,18 +177,15 @@ void ImagePreparer::sendNotes( sal_uInt32 aSlideNumber )
         return;
 
     // Start the writing
-    OStringBuffer aBuffer;
-
-    aBuffer.append( "slide_notes\n" );
-
-    aBuffer.append( static_cast<sal_Int32>(aSlideNumber) );
-    aBuffer.append( "\n" );
-
-    aBuffer.append( "<html><body>" );
-    aBuffer.append( aNotes );
-    aBuffer.append( "</body></html>" );
-    aBuffer.append( "\n\n" );
-    pTransmitter->addMessage( aBuffer.makeStringAndClear(),
+    OString aBuffer =
+        "slide_notes\n" +
+        OString::number( static_cast<sal_Int32>(aSlideNumber) ) +
+        "\n"
+        "<html><body>" +
+        aNotes +
+        "</body></html>"
+        "\n\n";
+    pTransmitter->addMessage( aBuffer,
         Transmitter::PRIORITY_LOW );
 }
 
@@ -222,17 +213,16 @@ OString ImagePreparer::prepareNotes( sal_uInt32 aSlideNumber )
     static const OUString sTextShapeName (
         "com.sun.star.drawing.TextShape" );
 
-    uno::Reference<container::XIndexAccess> xIndexAccess ( aNotesPage, UNO_QUERY);
-    if (xIndexAccess.is())
+    if (aNotesPage.is())
     {
 
         // Iterate over all shapes and find the one that holds the text.
-        sal_Int32 nCount (xIndexAccess->getCount());
+        sal_Int32 nCount (aNotesPage->getCount());
         for (sal_Int32 nIndex=0; nIndex<nCount; ++nIndex)
         {
 
             uno::Reference<lang::XServiceName> xServiceName (
-                xIndexAccess->getByIndex(nIndex), UNO_QUERY);
+                aNotesPage->getByIndex(nIndex), UNO_QUERY);
             if (xServiceName.is()
                 && xServiceName->getServiceName() == sNotesShapeName)
             {
@@ -246,14 +236,14 @@ OString ImagePreparer::prepareNotes( sal_uInt32 aSlideNumber )
             else
             {
                 uno::Reference<drawing::XShapeDescriptor> xShapeDescriptor (
-                    xIndexAccess->getByIndex(nIndex), UNO_QUERY);
+                    aNotesPage->getByIndex(nIndex), UNO_QUERY);
                 if (xShapeDescriptor.is())
                 {
                     OUString sType (xShapeDescriptor->getShapeType());
                     if (sType == sNotesShapeName || sType == sTextShapeName)
                     {
                         uno::Reference<text::XTextRange> xText (
-                            xIndexAccess->getByIndex(nIndex), UNO_QUERY);
+                            aNotesPage->getByIndex(nIndex), UNO_QUERY);
                         if (xText.is())
                         {
                             aRet.append(xText->getString());
